Fix connecting to peer bug
Previously, when connecting to a new peer in which a previous connection attempt was made to another peer that doesnt respond by either accepting or declining the connection request was not possible. Peer.close method should be called first before calling the connect method again.

It is now fixed.

Connecting to ownself also triggers the onpeerdecline listener.